Split the components.
This makes the install a lot easier to automate without any inputs. 
The Component names have to be retranslated, it also includes the Enhanced Edition file that were missing in the mod. 
The new .bif file names start with my(Jarno Mikkola/The Imp) and Lollorian's prefixes, which we give a full approval to utilize in .bif files to this mod.